Remove Program from VecOps in Rust (#985)
