+++ wrappers/rust/icicle-core/src/vec_ops/mod.rs
@@ -1,4 +1,3 @@
-use crate::program::Program;
 use crate::traits::FieldImpl;
 use icicle_runtime::{
     config::ConfigExtension, errors::eIcicleError, memory::HostOrDeviceSlice, stream::IcicleStreamHandle,
@@ -138,17 +137,6 @@ pub trait VecOps<F> {
         cfg: &VecOpsConfig,
         output: &mut (impl HostOrDeviceSlice<F> + ?Sized),
     ) -> Result<(), eIcicleError>;
-
-    fn execute_program<Prog, Data>(
-        data: &mut Vec<&Data>,
-        program: &Prog,
-        cfg: &VecOpsConfig,
-    ) -> Result<(), eIcicleError>
-    where
-        F: FieldImpl,
-        <F as FieldImpl>::Config: VecOps<F>,
-        Data: HostOrDeviceSlice<F> + ?Sized,
-        Prog: Program<F>;
 }
 
 #[doc(hidden)]
@@ -268,36 +256,6 @@ fn check_vec_ops_args_slice<F>(
     }
     let batch_size = output.len() / size_out as usize;
     setup_config(input, input, output, cfg, batch_size)
-}
-
-fn check_execute_program<F, Data>(data: &Vec<&Data>, cfg: &VecOpsConfig) -> VecOpsConfig
-where
-    F: FieldImpl,
-    <F as FieldImpl>::Config: VecOps<F>,
-    Data: HostOrDeviceSlice<F> + ?Sized,
-{
-    // All parameters' config should match so each one is compared to the first one
-    let nof_iterations = data[0].len();
-    let is_on_device = data[0].is_on_device();
-
-    for i in 1..data.len() {
-        if data[i].len() != nof_iterations {
-            panic!(
-                "First parameter length ({}) and parameter[{}] length do not match",
-                nof_iterations,
-                data[i].len()
-            );
-        }
-        if data[i].is_on_device() != is_on_device {
-            panic!(
-                "First parameter length ({}) and parameter[{}] length ({}) do not match",
-                nof_iterations,
-                i,
-                data[i].len()
-            );
-        }
-    }
-    setup_config(data[0], data[0], data[0], cfg, 1)
 }
 
 /// Modify VecopsConfig according to the given vectors
@@ -548,21 +506,6 @@ where
     <<F as FieldImpl>::Config as VecOps<F>>::slice(input, offset, stride, size_in, size_out, &cfg, output)
 }
 
-pub fn execute_program<F, Prog, Data>(
-    data: &mut Vec<&Data>,
-    program: &Prog,
-    cfg: &VecOpsConfig,
-) -> Result<(), eIcicleError>
-where
-    F: FieldImpl,
-    <F as FieldImpl>::Config: VecOps<F>,
-    Data: HostOrDeviceSlice<F> + ?Sized,
-    Prog: Program<F>,
-{
-    let cfg = check_execute_program(&data, cfg);
-    <<F as FieldImpl>::Config as VecOps<F>>::execute_program(data, program, &cfg)
-}
-
 #[macro_export]
 macro_rules! impl_vec_ops_field {
     (
@@ -573,8 +516,6 @@ macro_rules! impl_vec_ops_field {
     ) => {
         mod $field_prefix_ident {
             use crate::vec_ops::{$field, HostOrDeviceSlice};
-            use icicle_core::program::{Program, ProgramHandle};
-            use icicle_core::symbol::Symbol;
             use icicle_core::vec_ops::VecOpsConfig;
             use icicle_runtime::errors::eIcicleError;
 
@@ -702,14 +643,6 @@ macro_rules! impl_vec_ops_field {
                     output: *mut $field,
                 ) -> eIcicleError;
 
-                #[link_name = concat!($field_prefix, "_execute_program")]
-                pub(crate) fn execute_program_ffi(
-                    data_ptr: *const *const $field,
-                    nof_params: u64,
-                    program: ProgramHandle,
-                    nof_iterations: u64,
-                    cfg: *const VecOpsConfig,
-                ) -> eIcicleError;
             }
         }
 
@@ -976,32 +909,6 @@ macro_rules! impl_vec_ops_field {
                     .wrap()
                 }
             }
-
-            fn execute_program<Prog, Data>(
-                data: &mut Vec<&Data>,
-                program: &Prog,
-                cfg: &VecOpsConfig,
-            ) -> Result<(), eIcicleError>
-            where
-                <$field as FieldImpl>::Config: VecOps<$field>,
-                Data: HostOrDeviceSlice<$field> + ?Sized,
-                Prog: Program<$field>,
-            {
-                unsafe {
-                    let data_vec: Vec<*const $field> = data
-                        .iter()
-                        .map(|s| s.as_ptr())
-                        .collect();
-                    $field_prefix_ident::execute_program_ffi(
-                        data_vec.as_ptr(),
-                        data.len() as u64,
-                        program.handle(),
-                        data[0].len() as u64,
-                        cfg as *const VecOpsConfig,
-                    )
-                    .wrap()
-                }
-            }
         }
     };
 }
@@ -1063,7 +970,6 @@ macro_rules! impl_vec_ops_tests {
     ) => {
         pub(crate) mod test_vecops {
             use super::*;
-            use crate::program::$field_prefix_ident::{FieldProgram, FieldReturningValueProgram};
             use icicle_runtime::test_utilities;
             use icicle_runtime::{device::Device, runtime};
             use std::sync::Once;
@@ -1172,24 +1078,6 @@ macro_rules! impl_vec_ops_tests {
             pub fn test_slice() {
                 initialize();
                 check_slice::<$field>()
-            }
-
-            #[test]
-            pub fn test_program() {
-                initialize();
-                test_utilities::test_set_main_device();
-                check_program::<$field, FieldProgram>();
-                test_utilities::test_set_ref_device();
-                check_program::<$field, FieldProgram>()
-            }
-
-            #[test]
-            pub fn test_predefined_program() {
-                initialize();
-                test_utilities::test_set_main_device();
-                check_predefined_program::<$field, FieldProgram>();
-                test_utilities::test_set_ref_device();
-                check_predefined_program::<$field, FieldProgram>()
             }
         }
     };
