+++ wrappers/rust/icicle-core/src/vec_ops/tests.rs
@@ -6,9 +6,9 @@ use crate::symbol::Symbol;
 use crate::traits::{Arithmetic, GenerateRandom};
 use crate::vec_ops::poly_vecops::{polyvec_add, polyvec_mul, polyvec_mul_by_scalar, polyvec_sub, polyvec_sum_reduce};
 use crate::vec_ops::{
-    accumulate_scalars, add_scalars, bit_reverse, bit_reverse_inplace, div_scalars, execute_program, inv_scalars,
-    mixed_mul_scalars, mul_scalars, product_scalars, scalar_add, scalar_mul, scalar_sub, slice, sub_scalars,
-    sum_scalars, transpose_matrix, FieldImpl, MixedVecOps, VecOps, VecOpsConfig,
+    accumulate_scalars, add_scalars, bit_reverse, bit_reverse_inplace, div_scalars, inv_scalars, mixed_mul_scalars,
+    mul_scalars, product_scalars, scalar_add, scalar_mul, scalar_sub, slice, sub_scalars, sum_scalars,
+    transpose_matrix, FieldImpl, MixedVecOps, VecOps, VecOpsConfig,
 };
 use icicle_runtime::device::Device;
 use icicle_runtime::memory::{DeviceVec, HostOrDeviceSlice, HostSlice};
@@ -473,129 +473,6 @@ where
         .copy_to_host(HostSlice::from_mut_slice(&mut result_host[..]))
         .unwrap();
     assert_eq!(input.as_slice(), result_host.as_slice());
-}
-
-pub fn check_program<F, Prog>()
-where
-    F: FieldImpl,
-    <F as FieldImpl>::Config: VecOps<F> + GenerateRandom<F> + FieldArithmetic<F>,
-    Prog: Program<F>,
-{
-    let example_lambda = |vars: &mut Vec<Prog::ProgSymbol>| {
-        let a = vars[0]; // Shallow copies pointing to the same memory in the backend
-        let b = vars[1];
-        let c = vars[2];
-        let d = vars[3];
-
-        vars[4] = d * (a * b - c) + F::from_u32(9);
-        vars[5] = a * b - c.inverse();
-        vars[6] += a * b - c.inverse();
-        vars[3] = (vars[0] + vars[1]) * F::from_u32(2); // all variables can be both inputs and outputs
-    };
-
-    const TEST_SIZE: usize = 1 << 10;
-    let a = F::Config::generate_random(TEST_SIZE);
-    let b = F::Config::generate_random(TEST_SIZE);
-    let c = F::Config::generate_random(TEST_SIZE);
-    let eq = F::Config::generate_random(TEST_SIZE);
-    let var4 = vec![F::zero(); TEST_SIZE];
-    let var5 = vec![F::zero(); TEST_SIZE];
-    let var6 = vec![F::zero(); TEST_SIZE];
-    let a_slice = HostSlice::from_slice(&a);
-    let b_slice = HostSlice::from_slice(&b);
-    let c_slice = HostSlice::from_slice(&c);
-    let eq_slice = HostSlice::from_slice(&eq);
-    let var4_slice = HostSlice::from_slice(&var4);
-    let var5_slice = HostSlice::from_slice(&var5);
-    let var6_slice = HostSlice::from_slice(&var6);
-    let mut parameters = vec![a_slice, b_slice, c_slice, eq_slice, var4_slice, var5_slice, var6_slice];
-
-    let program = Prog::new(example_lambda, 7).unwrap();
-
-    let cfg = VecOpsConfig::default();
-    execute_program(&mut parameters, &program, &cfg).expect("Program Failed");
-
-    for i in 0..TEST_SIZE {
-        let a = a[i];
-        let b = b[i];
-        let c = c[i];
-        let eq = eq[i];
-        let var3 = parameters[3][i];
-        let var4 = parameters[4][i];
-        let var5 = parameters[5][i];
-        let var6 = parameters[6][i];
-        assert_eq!(
-            var3,
-            <<F as FieldImpl>::Config as FieldArithmetic<F>>::mul(
-                F::from_u32(2),
-                <<F as FieldImpl>::Config as FieldArithmetic<F>>::add(a, b)
-            )
-        );
-        assert_eq!(
-            var4,
-            <<F as FieldImpl>::Config as FieldArithmetic<F>>::add(
-                F::from_u32(9),
-                <<F as FieldImpl>::Config as FieldArithmetic<F>>::mul(
-                    eq,
-                    <<F as FieldImpl>::Config as FieldArithmetic<F>>::sub(
-                        <<F as FieldImpl>::Config as FieldArithmetic<F>>::mul(a, b),
-                        c
-                    )
-                )
-            )
-        );
-        assert_eq!(
-            var5,
-            <<F as FieldImpl>::Config as FieldArithmetic<F>>::sub(
-                <<F as FieldImpl>::Config as FieldArithmetic<F>>::mul(a, b),
-                <<F as FieldImpl>::Config as FieldArithmetic<F>>::inv(c)
-            )
-        );
-        assert_eq!(var6, var5);
-    }
-}
-
-pub fn check_predefined_program<F, Prog>()
-where
-    F: FieldImpl,
-    <F as FieldImpl>::Config: VecOps<F> + GenerateRandom<F> + FieldArithmetic<F>,
-    Prog: Program<F>,
-{
-    const TEST_SIZE: usize = 1 << 10;
-    let a = F::Config::generate_random(TEST_SIZE);
-    let b = F::Config::generate_random(TEST_SIZE);
-    let c = F::Config::generate_random(TEST_SIZE);
-    let eq = F::Config::generate_random(TEST_SIZE);
-    let var4 = vec![F::zero(); TEST_SIZE];
-    let a_slice = HostSlice::from_slice(&a);
-    let b_slice = HostSlice::from_slice(&b);
-    let c_slice = HostSlice::from_slice(&c);
-    let eq_slice = HostSlice::from_slice(&eq);
-    let var4_slice = HostSlice::from_slice(&var4);
-    let mut parameters = vec![a_slice, b_slice, c_slice, eq_slice, var4_slice];
-
-    let program = Prog::new_predefined(PreDefinedProgram::EQtimesABminusC).unwrap();
-
-    let cfg = VecOpsConfig::default();
-    execute_program(&mut parameters, &program, &cfg).expect("Program Failed");
-
-    for i in 0..TEST_SIZE {
-        let a = parameters[0][i];
-        let b = parameters[1][i];
-        let c = parameters[2][i];
-        let eq = parameters[3][i];
-        let var4 = parameters[4][i];
-        assert_eq!(
-            var4,
-            <<F as FieldImpl>::Config as FieldArithmetic<F>>::mul(
-                eq,
-                <<F as FieldImpl>::Config as FieldArithmetic<F>>::sub(
-                    <<F as FieldImpl>::Config as FieldArithmetic<F>>::mul(a, b),
-                    c
-                )
-            )
-        );
-    }
 }
 
 pub fn check_vec_ops_mixed_scalars_mul<F: FieldImpl, T: FieldImpl>(test_size: usize)
