+++ wrappers/rust/icicle-curves/icicle-bw6-761/src/program/mod.rs
@@ -3,3 +3,9 @@ use crate::curve::ScalarField;
 use icicle_core::impl_program_field;
 
 impl_program_field!("bw6_761", bw6_761, ScalarField, ScalarCfg);
+
+#[cfg(test)]
+use icicle_core::impl_program_tests;
+
+#[cfg(test)]
+impl_program_tests!(bw6_761, ScalarField);
