+++ wrappers/rust/icicle-curves/icicle-bw6-761/src/vec_ops/mod.rs
@@ -5,9 +5,6 @@ use icicle_core::{
 };
 use icicle_runtime::{errors::eIcicleError, memory::HostOrDeviceSlice};
 
-use icicle_core::program::Program;
-use icicle_core::traits::FieldImpl;
-
 impl_vec_ops_field!("bw6_761", bw6_761, ScalarField, ScalarCfg);
 
 #[cfg(test)]
