+++ wrappers/rust/icicle-fields/icicle-babybear/src/program/mod.rs
@@ -4,3 +4,16 @@ use icicle_core::impl_program_field;
 
 impl_program_field!("babybear", babybear, ScalarField, ScalarCfg);
 impl_program_field!("babybear_extension", babybear_extension, ExtensionField, ExtensionCfg);
+
+#[cfg(test)]
+pub(crate) mod tests {
+    use crate::field::{ExtensionField, ScalarField};
+    use icicle_core::impl_program_tests;
+
+    impl_program_tests!(babybear, ScalarField);
+
+    mod extension {
+        use super::*;
+        impl_program_tests!(babybear_extension, ExtensionField);
+    }
+}
