+++ wrappers/rust/icicle-fields/icicle-babybear/src/vec_ops/mod.rs
@@ -5,9 +5,6 @@ use icicle_core::{impl_vec_ops_field, impl_vec_ops_mixed_field};
 use icicle_runtime::errors::eIcicleError;
 use icicle_runtime::memory::HostOrDeviceSlice;
 
-use icicle_core::program::Program;
-use icicle_core::traits::FieldImpl;
-
 impl_vec_ops_field!("babybear", babybear, ScalarField, ScalarCfg);
 impl_vec_ops_field!("babybear_extension", babybear_extension, ExtensionField, ExtensionCfg);
 impl_vec_ops_mixed_field!(
