+++ wrappers/rust/icicle-fields/icicle-goldilocks/src/vec_ops/mod.rs
@@ -5,9 +5,6 @@ use icicle_core::vec_ops::{VecOps, VecOpsConfig};
 use icicle_runtime::errors::eIcicleError;
 use icicle_runtime::memory::HostOrDeviceSlice;
 
-use icicle_core::program::Program;
-use icicle_core::traits::FieldImpl;
-
 impl_vec_ops_field!("goldilocks", goldilocks, ScalarField, ScalarCfg);
 
 #[cfg(test)]
