+++ wrappers/rust/icicle-fields/icicle-koalabear/src/vec_ops/mod.rs
@@ -5,9 +5,6 @@ use icicle_core::{impl_vec_ops_field, impl_vec_ops_mixed_field};
 use icicle_runtime::errors::eIcicleError;
 use icicle_runtime::memory::HostOrDeviceSlice;
 
-use icicle_core::program::Program;
-use icicle_core::traits::FieldImpl;
-
 impl_vec_ops_field!("koalabear", koalabear, ScalarField, ScalarCfg);
 impl_vec_ops_field!("koalabear_extension", koalabear_extension, ExtensionField, ExtensionCfg);
 impl_vec_ops_mixed_field!(
