+++ wrappers/rust/icicle-fields/icicle-m31/src/program/mod.rs
@@ -4,3 +4,16 @@ use icicle_core::impl_program_field;
 
 impl_program_field!("m31", m31, ScalarField, ScalarCfg);
 impl_program_field!("m31_extension", m31_extension, ExtensionField, ExtensionCfg);
+
+#[cfg(test)]
+pub(crate) mod tests {
+    use crate::field::{ExtensionField, ScalarField};
+    use icicle_core::impl_program_tests;
+
+    impl_program_tests!(m31, ScalarField);
+
+    mod extension {
+        use super::*;
+        impl_program_tests!(m31_extension, ExtensionField);
+    }
+}
