+++ wrappers/rust/icicle-fields/icicle-m31/src/vec_ops/mod.rs
@@ -5,9 +5,6 @@ use icicle_core::{impl_vec_ops_field, impl_vec_ops_mixed_field};
 use icicle_runtime::errors::eIcicleError;
 use icicle_runtime::memory::HostOrDeviceSlice;
 
-use icicle_core::program::Program;
-use icicle_core::traits::FieldImpl;
-
 impl_vec_ops_field!("m31", m31, ScalarField, ScalarCfg);
 impl_vec_ops_field!("m31_extension", m31_extension, ExtensionField, ExtensionCfg);
 impl_vec_ops_mixed_field!("m31_extension", m31_mixed, ExtensionField, ScalarField, ExtensionCfg);
