+++ wrappers/rust/icicle-fields/icicle-stark252/src/program/mod.rs
@@ -3,3 +3,9 @@ use crate::field::ScalarField;
 use icicle_core::impl_program_field;
 
 impl_program_field!("stark252", stark252, ScalarField, ScalarCfg);
+
+#[cfg(test)]
+use icicle_core::impl_program_tests;
+
+#[cfg(test)]
+impl_program_tests!(stark252, ScalarField);
