+++ wrappers/rust/icicle-fields/icicle-stark252/src/vec_ops/mod.rs
@@ -5,9 +5,6 @@ use icicle_core::vec_ops::{VecOps, VecOpsConfig};
 use icicle_runtime::errors::eIcicleError;
 use icicle_runtime::memory::HostOrDeviceSlice;
 
-use icicle_core::program::Program;
-use icicle_core::traits::FieldImpl;
-
 impl_vec_ops_field!("stark252", stark252, ScalarField, ScalarCfg);
 
 #[cfg(test)]
